+++ feature-userInfo/src/main/java/com/hmoa/feature_userinfo/paging/OrderPagingSource.kt
@@ -1,26 +1,26 @@
-package com.hmoa.feature_userinfo
+package com.hmoa.feature_userinfo.paging
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.hmoa.core_domain.repository.CommunityRepository
-import com.hmoa.core_model.response.CommunityByCategoryResponseDto
+import com.hmoa.core_domain.repository.MemberRepository
+import com.hmoa.core_model.response.OrderRecordDto
 
-class CommunityPagingSource(
-    private val communityRepository: CommunityRepository
-) : PagingSource<Int, CommunityByCategoryResponseDto>() {
+class OrderPagingSource(
+    private val memberRepository: MemberRepository
+) : PagingSource<Int, OrderRecordDto>() {
 
     private var totalCount = 0
     private var cursor = 0
 
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, CommunityByCategoryResponseDto> {
+    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, OrderRecordDto> {
         val pageNumber = params.key ?: 0
-        val response = communityRepository.getMyCommunities(cursor)
+        val response = memberRepository.getOrder(cursor)
         if (response.errorMessage != null) {
             return LoadResult.Error(Exception(response.errorMessage!!.message))
         }
         val result = response.data!!
         totalCount = result.data.size
-        cursor = if (totalCount == 0) 0 else result.data.last().communityId
+        cursor = if (totalCount == 0) 0 else result.data.last().orderId
         val prevKey = if (pageNumber > 0) pageNumber - 1 else null
         val nextKey = if (result.lastPage) null else pageNumber + 1
 
@@ -31,7 +31,7 @@ class CommunityPagingSource(
         )
     }
 
-    override fun getRefreshKey(state: PagingState<Int, CommunityByCategoryResponseDto>): Int? {
+    override fun getRefreshKey(state: PagingState<Int, OrderRecordDto>): Int? {
         return state.anchorPosition?.let {
             state.closestPageToPosition(it)?.prevKey?.plus(1)
                 ?: state.closestPageToPosition(it)?.nextKey?.minus(1)
